+++ DolbomMon/src/main/java/com/dolbommon/dbmon/search/TeacherVO.java
@@ -21,6 +21,8 @@ public class TeacherVO {
 	private MultipartFile filename;
 	private int hit;
 	private String last_edit;
+	private int headcount;
+	private String discussion;
 	
 	public MultipartFile getFilename() {
 		return filename;
@@ -130,7 +132,18 @@ public class TeacherVO {
 	public void setLast_edit(String last_edit) {
 		this.last_edit = last_edit;
 	}
-	
+	public int getHeadcount() {
+		return headcount;
+	}
+	public void setHeadcount(int headcount) {
+		this.headcount = headcount;
+	}
+	public String getDiscussion() {
+		return discussion;
+	}
+	public void setDiscussion(String discussion) {
+		this.discussion = discussion;
+	}
